% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constr}
\alias{constr}
\title{Generate Constraint Matrix for Quadratic Programming}
\usage{
constr(index, gamma = 4)
}
\arguments{
\item{index}{Grouping index}

\item{gamma}{Sensitivity parameter}
}
\value{
Constraint matrix A
}
\description{
Generate Constraint Matrix for Quadratic Programming
}
