% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_predict}
\alias{get_predict}
\title{Prediction Wrapper Function}
\usage{
get_predict(Y, x, x.2, model = "rf", ...)
}
\arguments{
\item{Y}{Response variable}

\item{x}{Predictors for training}

\item{x.2}{Predictors for prediction}

\item{model}{Model type}

\item{...}{Additional parameters}
}
\value{
Predicted values
}
\description{
Prediction Wrapper Function
}
