% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Q}
\alias{get_Q}
\title{Generate Quadratic Coefficient Matrix for Quadratic Programming}
\usage{
get_Q(vec, Z, index, c)
}
\arguments{
\item{vec}{Numeric vector}

\item{Z}{Treatment assignment vector}

\item{index}{Grouping index}

\item{c}{Chi-squared critical value}
}
\value{
Symmetric matrix Q
}
\description{
Generate Quadratic Coefficient Matrix for Quadratic Programming
}
