% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_w}
\alias{get_w}
\title{Calculate Linear Coefficients for Quadratic Programming}
\usage{
get_w(vec, c, tstar, ns1, ns2)
}
\arguments{
\item{vec}{Numeric vector}

\item{c}{Chi-squared critical value}

\item{tstar}{Test statistic}

\item{ns1}{Groups for one-to-one matching}

\item{ns2}{Groups for many-to-one matching}
}
\value{
Linear coefficients w
}
\description{
Calculate Linear Coefficients for Quadratic Programming
}
