% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m.logistic}
\alias{m.logistic}
\title{Logistic Regression Prediction Function}
\usage{
m.logistic(Y, x, x.2, ...)
}
\arguments{
\item{Y}{Response variable for training}

\item{x}{Predictors for training}

\item{x.2}{Predictors for prediction}

\item{...}{Additional GLM parameters}
}
\value{
A vector of predicted probabilities
}
\description{
Logistic Regression Prediction Function
}
