% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m.rf}
\alias{m.rf}
\title{Random Forest Prediction Function}
\usage{
m.rf(Y, x, x.2, ...)
}
\arguments{
\item{Y}{Response variable for training}

\item{x}{Predictors for training}

\item{x.2}{Predictors for prediction}

\item{...}{Additional RF parameters}
}
\value{
A vector of predicted probabilities
}
\description{
Random Forest Prediction Function
}
