% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m.svm}
\alias{m.svm}
\title{Support Vector Machine (SVM) Prediction Function}
\usage{
m.svm(Y, x, x.2, ...)
}
\arguments{
\item{Y}{Response variable for training}

\item{x}{Predictors for training}

\item{x.2}{Predictors for prediction}

\item{...}{Additional SVM parameters}
}
\value{
A vector of predicted values
}
\description{
Support Vector Machine (SVM) Prediction Function
}
