% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m.xgboost}
\alias{m.xgboost}
\title{XGBoost Prediction Function}
\usage{
m.xgboost(Y, x, x.2, ...)
}
\arguments{
\item{Y}{Numeric vector of response variable for the training data (1 for treated, 0 for control)}

\item{x}{Data frame or matrix of predictor variables for training}

\item{x.2}{Data frame or matrix of predictor variables for prediction}

\item{...}{Additional parameters to pass to the XGBoost model}
}
\value{
A numeric vector of predicted values (probabilities)
}
\description{
XGBoost Prediction Function
}
