% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permutation_split}
\alias{permutation_split}
\title{Permutation-Based Group Splitting Function}
\usage{
permutation_split(treated_lst, control_lst)
}
\arguments{
\item{treated_lst}{List of indices for the treated group}

\item{control_lst}{List of indices for the control group}
}
\value{
A list containing two grouped indices, each as a combination of treated and control indices
}
\description{
This function performs a permutation-based split of treated and control groups, shuffling and dividing
them into two balanced groups for further analysis. It ensures that single-element indices are grouped first.
}
